feat: allow different modes for username mapping
This introduces the `usernameMapper.mode` config key, which can be `plain` or `hmac-sha256`,
and controls how the `UsernameMapper` maps the username to the matrix localpart.

It can be used to turn "off" the hashing of the username, if the configured
password-provider already provides a username which meets the criteria of a
persistent ID (and is, for example, resistent to name or nickname changes).